electrum: clean up tests, and add watcher docker tests
